+++ ext/pkce_oauth/src/lib.rs
@@ -1,13 +1,19 @@
+use base64_url::escape;
 use magnus::{function, prelude::*, Error, Ruby};
 use rand::{thread_rng, Rng};
-use rand::distributions::Alphanumeric;
+//use rand::distributions::Alphanumeric;
 use sha256::digest;
 
 fn generate_limited_code_verifier(length: usize) -> String {
-  thread_rng()
-    .sample_iter(&Alphanumeric)
-    .take(length)
-    .map(char::from)
+  const CHARSET: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZ\
+                            abcdefghijklmnopqrstuvwxyz\
+                            0123456789-_";
+  let mut rng = thread_rng();
+  (0..length)
+    .map(|_| {
+      let index = rng.gen_range(0..CHARSET.len());
+      CHARSET[index] as char
+    })
     .collect()
 }
 
@@ -16,7 +22,7 @@ fn generate_code_verifier() -> String {
 }
 
 fn generate_code_challenge(code_verifier: String) -> String {
-  digest(code_verifier)
+  escape(&digest(code_verifier)).to_string()
 }
 
 #[magnus::init]