+++ lib/pkce_oauth/challenger.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module PkceOauth
+  class Challenger
+    def call(code_verifier:)
+      PkceOauth.generate_code_challenge(code_verifier)
+    end
+  end
+end