+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-## [Unreleased]
+## [0.2.0] - 2024-12-16
+### Added
+- generating random key with charset
+- url escaping for code challenge
+
+### Modified
+- readme
 
 ## [0.1.0] - 2024-12-16